v1.0.10
## Fixes

- fixed `COMPARISON_CHECKS` assertions' `CompareTo` calls to allow for `T` being of the type `byte`, as `byte.CompareTo` does not necessarily only return 0, 1 and -1

+++ Unity/Editor/SafetyCheckSettingsWindow.cs
@@ -20,7 +20,7 @@ namespace DevTools.Unity.Editor
         private bool firstLoad;
         private bool anythingChanged;
 
-        private Dictionary<Assert.GroupAttribute, ulong> methodCallCounts = null;
+        private Dictionary<Assert.GroupAttribute, uint> methodCallCounts = null;
         private Dictionary<Assert.GroupAttribute, bool> definesToCheckMarks;
 
 
@@ -48,7 +48,7 @@ namespace DevTools.Unity.Editor
             {
                 ulong count = 0;
 
-                foreach (KeyValuePair<Assert.GroupAttribute, ulong> assertion in methodCallCounts)
+                foreach (KeyValuePair<Assert.GroupAttribute, uint> assertion in methodCallCounts)
                 {
                     count += assertion.Value;
                 }
